% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_spelling_function}
\alias{test_spelling_function}
\title{Only simple function}
\usage{
test_spelling_function(x)
}
\arguments{
\item{x}{This text contains the weird word Xargrds}
}
\value{
This is good text.
}
\description{
Only simple function
}
